WildLens-Model/requirements.txt
# NOTE: PyTorch is NOT explicitly pinned here because CUDA builds must match your NVIDIA driver.
# After installing these packages, run the CUDA installer helper:
#   python .\WildLens-Model\scripts\install_torch_cuda.py --yes
# This will uninstall any CPU-only torch and install a matching CUDA wheel.
# If you prefer CPU-only environments, skip the helper and let pip install CPU torch implicitly via ultralytics.

ultralytics==8.3.50
roboflow>=1.1.25
onnx>=1.16.0
# Use GPU-enabled ONNX Runtime on machines with CUDA. If you need CPU-only, install onnxruntime instead.
onnxruntime-gpu>=1.18.0
opencv-python>=4.9.0.80
PyYAML>=6.0.1
numpy>=1.24.0
